feat: complete service layer modularization (Phase 7 - issue #84)
Remove old monolithic service.py implementation and complete migration to
modular ServiceManager architecture. All functionality preserved while
improving maintainability and testability.

Changes:
- Remove src/minerva/service.py and tests/test_service.py
- Update server.py to use ServiceManager directly
- Fix type annotations in test files
- Update comprehensive documentation across docs/ and .github/instructions/
- Maintain all existing functionality with 410 passing tests
